requirements.txt
pyyaml
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gymnasium==1.0.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
minigrid==3.0.0
mpmath==1.3.0
networkx==3.6
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pygame==2.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
stable_baselines3==2.4.1
sympy==1.14.0
torch==2.2.2
typing_extensions==4.15.0
tzdata==2025.2
